backend/python/requirements.txt
tensorflow; python_version < "3.13"
numpy
flask
pillow
joblib
scikit-learn
